+++ Cargo.toml
@@ -27,8 +27,7 @@ regex = "1"
 reqwest = "0"
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1" }
-# tantivy = "0.22"
-tantivy = { path = "../tantivy" }
+tantivy = "0.22"
 tokio = { version = "1", features = ["full"] }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 zstd-sys = "=2.0.9"
